adding apis for inserting,getting and deleting a book

+++ src/main/java/com/bookrental/Services/BookService.java
@@ -1,11 +1,15 @@
 package com.bookrental.Services;
 
+import com.bookrental.Models.Author;
 import com.bookrental.Models.Book;
+import com.bookrental.Repositories.AuthorRepository;
 import com.bookrental.Repositories.BookRepository;
+import com.bookrental.Repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 
 @Service
@@ -13,6 +17,10 @@ public class BookService {
 
     @Autowired
     BookRepository bookRepository;
+    @Autowired
+    AuthorRepository authorRepository;
+    @Autowired
+    CategoryRepository categoryRepository;
 
     public BookService() {
     }
@@ -27,14 +35,26 @@ public class BookService {
         return bookRepository.findOne(id);
     }
 
-    public boolean add(Book book) {
+    public Boolean add(Book book) {
         book.setCopy();
-        bookRepository.save(book);
-        return true;
+        book.getAuthors().forEach(author -> {
+            if(authorRepository.findByName(author.getname()) != null)
+                author.setId(authorRepository.findByName(author.getname()).getId());
+        });
+        book.getCategories().forEach(category -> {
+            if(categoryRepository.findByName(category.getName()) != null)
+                category.setId(categoryRepository.findByName(category.getName()).getId());
+        });
+        if(bookRepository.save(book)!=null)
+            return true;
+        return false;
     }
 
     public boolean delete(String id) {
-        bookRepository.delete(id);
-        return true;
+        if(bookRepository.findOne(id)!=null) {
+            bookRepository.delete(id);
+            return true;
+        }
+        return false;
     }
 }
